Added function 'str_single_spaces' in string manipulation functions in 'libmisc'. GPU pipeline debug info updates. Warning removed in 'm2s-opencl-kc' tool. Debug information of active mask stack updates.

+++ trunk/tools/m2s-opencl-kc/m2s-opencl-kc.c
@@ -119,7 +119,7 @@ void kernel_binary_analyze(char *file_name)
 
 	char file_name_dest[MAX_STRING_SIZE];
 	void *section_buf;
-	size_t section_size;
+	uint32_t section_size;
 	char *section_name;
 
 	/* Get file name prefix */
